All kinds of changes
Update UpdateApp to get the proxy login details according to the default
proxy.
Implement tabs  on the shell screen.
When re-showing the window, bring to front.
Add a todo.txt with all the things that I still want to do.
Automatically bring up the window when plugging in a device.